[DEV-4069] Missed import renames
