+++ usaspending_api/search/v2/views/spending_by_category_views/abstract_spending_by_agency.py
@@ -7,7 +7,7 @@ from typing import List
 from django.db.models import QuerySet, F
 
 from usaspending_api.search.helpers.spending_by_category_helpers import fetch_agency_tier_id_by_agency
-from usaspending_api.search.v2.views.spending_by_category_views.base_spending_by_category import (
+from usaspending_api.search.v2.views.spending_by_category_views.abstract_spending_by_category import (
     AbstractSpendingByCategoryViewSet,
 )
