+++ usaspending_api/search/v2/views/spending_by_category_views/spending_by_industry_codes.py
@@ -1,5 +1,5 @@
-from usaspending_api.search.v2.views.spending_by_category_views.base_spending_by_category import Category
-from usaspending_api.search.v2.views.spending_by_category_views.base_spending_by_industry_codes import (
+from usaspending_api.search.v2.views.spending_by_category_views.abstract_spending_by_category import Category
+from usaspending_api.search.v2.views.spending_by_category_views.abstract_spending_by_industry_codes import (
     IndustryCodeType,
     AbstractIndustryCodeViewSet,
 )
